Hook up Settings Service
SettingsService and UnitFormatter have been configured to update the UI via Combine when the user changes the settings from metric to imperial.

SettingsView has been updated to allow the user to change the Units in use.

UserDefaults used to store the settings selected by the user. Also added a check in for first launch and added a method to set up defaults on a first run of the app.